Rename useGetAccountName and getAccountName to useGetSafeName and getSafeName
